Demo: instancing with unique transforms
I want to create a demo where instancing with unique transforms are
applied and controlled by user settings.

- Transforms are now defined in the global set
- Frustum culling is now being done through a BVH
- Scales with large instance counts

+++ src/engine/platform/profiler/Profiler.cpp
@@ -109,7 +109,7 @@ float Profiler::endTimer() const {
 	LARGE_INTEGER now;
 	QueryPerformanceCounter(&now);
 	auto elapsedTicks = now.QuadPart - _startTimer.QuadPart;
-	return static_cast<float>(elapsedTicks / _qpcFreq.QuadPart);
+	return static_cast<float>(elapsedTicks) / static_cast<float>(_qpcFreq.QuadPart);
 }
 
 VkDeviceSize Profiler::GetTotalVRAMUsage(VkPhysicalDevice device, VmaAllocator allocator) {
